third_party/blink/web_tests/http/tests/devtools/a11y-axe-core/performance/performance-pane-a11y-test-expected.txt


Running: testDetailsView
Tests accessibility in performance Details view using the axe-core linter
aXe violations: []


Running: testBottomUpView
Tests accessibility in performance BottomUp view using the axe-core linter
aXe violations: []


Running: testCallTreeView
Tests accessibility in performance CallTree view using the axe-core linter
aXe violations: []
